% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{calc_dispersion}
\alias{calc_dispersion}
\title{Calculate gene dispersion of a matrix}
\usage{
calc_dispersion(obj, min_cells_detected = 1, min_exprs = 1,
  id_tag = "id", removeOutliers = TRUE)
}
\value{
a list containing 1) a table of dispersion values and 2) the function to fit
}
\description{
Many genes typically co-vary with one another, and so the dimensionality of the
data can be reduced with a wide variety of different algorithms. After calculating dispersion for
a using the \code{calc_dispersion} function, the 
\code{select_genes} function allows the user to identify a set of genes
that will be used in downstream dimensionality reduction methods.
}
\references{
Dispersion fitting taken from monocle2
}
