% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{center.and.scale.by.row}
\alias{center.and.scale.by.row}
\title{Center and scale by row}
\usage{
center.and.scale.by.row(dataset)
}
\arguments{
\item{dataset}{input matrix}
}
\description{
Just like it sounds but old school
}
