% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{convert_to_long}
\alias{convert_to_long}
\title{Converts to long data format}
\usage{
convert_to_long(object, coldata = "all",
  values_to = "NormalizedExpression", names_to = "sample")
}
\arguments{
\item{coldata}{a factor or dataframe of factors used to add information to long data table, the string "all" can be used if the class of the input
object is a DESeq2 dataset and all coldata information is to be copied to long data table}

\item{matrix}{Input object with unique rownames}
}
\description{
Will convert a data object (either DESeq2 object or a matrix) to a data table for further work.
}
