% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExport.R
\name{dataExport}
\alias{dataExport}
\title{Export data to Excel file}
\usage{
dataExport(
  object,
  filename,
  data_type = c("normalized_counts", "raw_counts", "vst_transformed"),
  unique_id_column = "id",
  symbol_name = "gene_short_name"
)
}
\arguments{
\item{object}{the object from which function will extract data.  Currently supported objects include:
1) DESeq2 object and 2) Bioconductor expression set.}

\item{filename}{filename of Excel file}

\item{data_type}{can be one of normalized, raw, transformed.}
}
\description{
This function takes data in various forms (DESeq, ExpressionSet and exports them to an Excel File)
}
\references{
Love, M.I., Huber, W., Anders, S. Moderated estimation of fold change and dispersion for
RNA-seq data with DESeq2 Genome Biology 15(12):550 (2014)

Orchestrating high-throughput genomic analysis with Bioconductor. W. Huber, V.J. Carey, 
R. Gentleman, ..., M. Morgan Nature Methods, 2015:12, 115.
}
