% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{plot_gene_dispersion}
\alias{plot_gene_dispersion}
\title{Plot dispersion}
\usage{
plot_gene_dispersion(obj)
}
\arguments{
\item{obj}{the output of the \code{calc_dispersion} function.}
}
\description{
Many genes typically co-vary with one another, and so the dimensionality of the
data can be reduced with a wide variety of different algorithms. After calculating dispersion for
a using the \code{calc_dispersion} function, the 
\code{select_genes} function allows the user to identify a set of genes
that will be used in downstream dimensionality reduction methods.
}
