% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGene}
\alias{plotGene}
\title{Plot genes from a matrix, DESeq2 object or limma object}
\usage{
plotGene(obj, ...)
}
\arguments{
\item{obj}{input object}

\item{gene}{gene input}

\item{intgroup}{factor to be used for grouping,}

\item{normalized}{normalize data (default is TRUE)}

\item{transform}{transform data (default is FALSE)}

\item{main}{title}

\item{xlab}{label for x axis of plot}

\item{returnData}{return data instead of plot}

\item{replaced}{= FALSE}

\item{theme_opts}{use theme_opts (default is T)}

\item{annotate_data}{default F}

\item{gene_short_name}{column containin gene symbol; default = "gene_short_name"}

\item{ENSID}{column containing gene id (ENSEMBL) default = "GENEID"}

\item{pc}{pseudocount}

\item{annotation_col}{the column of meta data to use for annotation (default ="sample")}

\item{point_size}{default 3}

\item{colors}{colors to use}
}
\description{
Plot genes from a matrix, DESeq2 object or limma object
}
