% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{remove_duplicated_rows}
\alias{remove_duplicated_rows}
\title{Removes data (collapses) with duplicated names}
\usage{
remove_duplicated_rows(matrix, labels)
}
\arguments{
\item{matrix}{Input object with unique rownames}

\item{labels}{the non-unique labels (i.e. Gene Symbol)}
}
\description{
Will collapse a matrix based on duplicated labels that have the greatest variance
}
