\name{seqGlue-package}
\alias{seqGlue-package}
\alias{seqGlue}
\docType{package}
\title{
  Makes sequencing data come together like glue
}
\description{
  More to come here
}
\details{
  Currently we have one function: makeGSEAfiles
}
\author{
Scott Furlan

Maintainer: Scott Furlan <scottfurlan@gmail.com>
}
\references{
  None yet
}
\keyword{ package }
\examples{
     writeGSEAfiles(dds, class_labels="group") 
}
