% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{writeGSEAfiles}
\alias{writeGSEAfiles}
\title{Write GSEA files}
\usage{
writeGSEAfiles(object, class_labels, gene_labels = "gene_short_name",
  gct_filename = "data.gct", cls_filename = "data.cls",
  normalized = FALSE)
}
\arguments{
\item{object}{the object from which function will extract data.  Currently supported objects include:
1) DESeq2 object and 2) Bioconductor expression set.}

\item{class_labels}{the meta data column name (as a string) whose data will be written into the .CLS file.}

\item{gene_labels}{the column name of the feature data that contains the desired name of .GCT rownames (ie genes; default is "gene_short_name").}

\item{gct_filename}{filename of the .GCT file.}

\item{cls_filename}{filename of the .CLS file.}
}
\description{
GSEA formats are cumbersome.  This function writes working .GCT and .CLS files for use in a variety of GSEA
software including but not limited to the desktop (Java) application and the java cli.
}
\references{
Love, M.I., Huber, W., Anders, S. Moderated estimation of fold change and dispersion for
RNA-seq data with DESeq2 Genome Biology 15(12):550 (2014)

Orchestrating high-throughput genomic analysis with Bioconductor. W. Huber, V.J. Carey, 
R. Gentleman, ..., M. Morgan Nature Methods, 2015:12, 115.
}
